Added updated code. Needs cleanup
