+++ gtas-parent/gtas-commons/src/main/java/gov/gtas/constant/EventReportConstants.java
@@ -1,0 +1,70 @@
+package gov.gtas.constant;
+
+public class EventReportConstants {
+	
+	/** TABLE COLUMN HEADER TEXT **/
+	
+	/** COVER PAGE TABLE COLUMN HEADER TEXT **/
+	final public static String COVER_PAGE_TABLE_COLUMN_LABEL_NAME = "Name";
+	final public static String COVER_PAGE_TABLE_COLUMN_LABEL_FLIGHT = "Flight";
+	final public static String COVER_PAGE_TABLE_COLUMN_LABEL_ORIGIN = "Origin";
+	final public static String COVER_PAGE_TABLE_COLUMN_LABEL_ETA = "ETA";
+	final public static String COVER_PAGE_TABLE_COLUMN_LABEL_ETD = "ETD";
+	final public static String COVER_PAGE_TABLE_COLUMN_LABEL_REPORT_DATE = "Report Date";
+	
+	/*PASSENGER TABLE COLUMN HEADER TEXT */
+
+	final public static String PASSENGER_TABLE_COLUMN_LABEL_NAME = "Name";
+	final public static String PASSENGER_TABLE_COLUMN_LABEL_DOB = "DOB";
+	final public static String PASSENGER_TABLE_COLUMN_LABEL_GENDER = "Gender";
+	final public static String PASSENGERP_TABLE_COLUMN_LABEL_NATIONALITY = "Nationality";
+	
+	
+	
+	/*FLIGHT TABLE COLUMN HEADER TEXT */
+	final public static String TRIP_TABLE_COLUMN_LABEL_FLIGHT_NUMBER = "Flight Number";
+	final public static String TRIP_TABLE_COLUMN_LABEL_CARRIER = "Carrier";
+	final public static String TRIP_TABLE_COLUMN_LABEL_ORIG_AIRPORT = "Origin Airport";
+	final public static String TRIP_TABLE_COLUMN_LABEL_DEST_AIRPORT = "Destination Airport";
+	final public static String TRIP_TABLE_COLUMN_LABEL_TRIP_TYPE = "Trip Type";
+	/*DOCUMENTS TABLE COLUMN HEADER TEXT */
+	
+	/*EVENT HIT INFORMATION COLUMN HEADER TEXT */
+	
+	/*HIT HISTORY COLUMN HEADER TEXT  */
+	
+	/*CO-TRAVELERS COLUMN HEADER TEXT */
+	
+	/*CREDIT CARS COLUMN HEADER TEXT */
+	
+	/*EMAIL ADDRESS TABLE COLUMN HEADER TEXT */
+	
+	/*PHONE NUMBER TABLE COLUMN HEADER TEXT */
+	
+	
+	/*FLIGHT HISTORY COLUMN HEADER TEXT */
+	
+	
+	/*EVENT NOTES COLUMN HEADER TEXT */
+	
+	
+	/*EVENT NOTES HISTORY*/
+	
+	/** TABLE HEADER TEXT **/
+	final public static String PASSENGER_TITLE = "Passenger";
+	final public static String REPORT_NAME = "GTAS Event Report";
+	final public static String PASSENGER_DETAIL_HEADER_TEXT = "Event ID";
+	final public static String PASSENGER_DETAIL_HEADER_TEXT_2 = "Highest Hit Severity";
+	final public static String DOCUMENTS_TITLE = "Documents";
+	final public static String EVENT_HIT_INFORMATION_TITLE = "Event Hit Information";
+	final public static String HIT_HISTORY_TITLE = "Hit History";
+	final public static String COTRAVELER_TITLE = "Co-travelers";
+	final public static String FLIGHT_HISTORY_TITLE = "Flight History";
+	final public static String CREDIT_CARD_TITLE = "Credit Cards";
+	final public static String EMAIL_TITLE = "Email Addresses";
+	final public static String PHONE_TITLE = "Phone Numbers";
+	final public static String EVENT_NOTES_TITLE = "Event Notes";
+	final public static String EVENT_NOTE_HISTORY_TITLE = "Event Notes History";
+	final public static String TRIP_TITLE = "Trip";
+
+}
