+++ gtas-parent/gtas-commons/src/main/java/gov/gtas/services/PassengerEventPdfReportService.java
@@ -1,0 +1,12 @@
+package gov.gtas.services;
+
+import gov.gtas.services.dto.PaxDetailPdfDocResponse;
+
+public interface PassengerEventPdfReportService {
+	
+	public PaxDetailPdfDocResponse createPassengerEventReport(Long paxId, Long flightId);
+	
+	
+	
+
+}
