--- gtas-parent/gtas-commons/src/main/java/gov/gtas/services/PassengerPdfService.java
@@ -1,10 +1,0 @@
-package gov.gtas.services;
-
-import gov.gtas.services.dto.PdfDocumentRequest;
-import gov.gtas.services.dto.PdfDocumentResponse;
-
-public interface PassengerPdfService < R extends PdfDocumentRequest, S extends PdfDocumentResponse> {
-	
-	public S createPaxDetailReport(R pdfDocumentRequest) throws Exception;
-
-}
